move device_communicators under distributed
