+++ vllm/distributed/communication_op.py
@@ -22,8 +22,8 @@ def tensor_model_parallel_all_reduce(input_: torch.Tensor) -> torch.Tensor:
     TLDR: always assume this function modifies its input, but use the return
     value as the output.
     """
-    from vllm.device_communicators import pynccl_utils
-    from vllm.device_communicators.custom_all_reduce import custom_all_reduce
+    from vllm.distributed.device_communicators import pynccl_utils
+    from vllm.distributed.device_communicators.custom_all_reduce import custom_all_reduce
 
     # Bypass the function if we are using only 1 GPU.
     if get_tensor_model_parallel_world_size() == 1:
